requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pypi_0
blas=1.0=mkl
blur-detector=0.0.6=pypi_0
brotli=1.0.9=h2bbff1b_8
brotli-bin=1.0.9=h2bbff1b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.6.2=h56e8100_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=py312h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py312hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=hd77b12b_0
fabio=2024.4.0=pypi_0
fastjsonschema=2.20.0=pypi_0
filelock=3.15.3=pypi_0
fonttools=4.51.0=py312h2bbff1b_0
freetype=2.12.1=ha860e81_0
fsspec=2024.6.0=pypi_0
h5py=3.11.0=pypi_0
hdf5plugin=4.4.0=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=73.1=h6c2663c_0
imageio=2.33.1=py312haa95532_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
intel-openmp=2021.4.0=pypi_0
ipykernel=6.29.4=pyh4bbf305_0
ipython=8.25.0=pyh7428d3b_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
jpeg=9e=h2bbff1b_1
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py312haa95532_0
kiwisolver=1.4.4=py312hd77b12b_0
krb5=1.20.1=h5b6d351_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libpng=1.6.39=h8cc25b3_0
libpq=12.17=h906ac69_0
libsodium=1.0.18=h8d14728_1
libtiff=4.5.1=hd77b12b_0
libwebp-base=1.3.2=h2bbff1b_0
lxml=5.2.2=pypi_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.5=pypi_0
matplotlib=3.8.4=py312haa95532_0
matplotlib-base=3.8.4=py312hc7c4135_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2021.4.0=pypi_0
mkl-service=2.4.0=py312h2bbff1b_1
mkl_fft=1.3.8=py312h2bbff1b_0
mkl_random=1.2.4=py312h59b6b97_0
mpmath=1.3.0=pypi_0
nbformat=5.10.4=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.3=pypi_0
numpy=1.26.4=py312hfd52020_0
numpy-base=1.26.4=py312h4dde369_0
opencv-python=4.10.0.84=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.14=h827c3e9_0
packaging=23.2=py312haa95532_0
pandas=2.2.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py312h2bbff1b_0
pip=24.0=py312haa95532_0
platformdirs=4.2.2=pyhd8ed1ab_0
plotly=5.22.0=pypi_0
ply=3.11=py312haa95532_1
prompt-toolkit=3.0.47=pyha770c72_0
psutil=5.9.0=py312h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pybind11-abi=5=hd3eb1b0_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.0.9=py312haa95532_0
pyqt=5.15.10=py312hd77b12b_0
pyqt5-sip=12.13.0=py312h2bbff1b_0
python=3.12.3=h1d929f7_1
python-dateutil=2.9.0post0=py312haa95532_2
pytz=2024.1=pypi_0
pywin32=305=py312h2bbff1b_0
pyzmq=25.1.2=py312hd77b12b_0
qt-main=5.15.2=h19c9488_10
referencing=0.35.1=pypi_0
rpds-py=0.18.1=pypi_0
scikit-image=0.24.0=pypi_0
scipy=1.13.1=py312hbb039d4_0
setuptools=69.5.1=py312haa95532_0
silx=2.1.0=pypi_0
sip=6.7.12=py312hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12.1=pypi_0
tbb=2021.12.0=pypi_0
tenacity=8.4.2=pypi_0
tifffile=2024.6.18=pypi_0
tk=8.6.14=h0416ee5_0
torch=2.3.1+cu118=pypi_0
torchaudio=2.3.1+cu118=pypi_0
torchvision=0.18.1=pypi_0
tornado=6.3.3=py312h2bbff1b_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.1=pypi_0
unicodedata2=15.1.0=py312h2bbff1b_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=py312haa95532_0
xz=5.4.6=h8cc25b3_1
zeromq=4.3.5=hd77b12b_0
zipp=3.19.2=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.5=hd43e919_2
